refactor(core/kit): move publishable embedded runbook to separate mod
